+++ src/app/(main-app)/fabric/new/page.tsx
@@ -1,0 +1,10 @@
+import { FabricNewForm } from '@/components/forms/fabric/new';
+const page = () => {
+  return (
+    <>
+      <FabricNewForm />
+    </>
+  );
+};
+
+export default page;